added hts in test
